% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outliers.R
\name{outlier}
\alias{outlier}
\title{Outlier Analysis}
\usage{
outlier(data)
}
\arguments{
\item{data}{= dataframe}
}
\description{
Takes dataframe and shows the outliers and their log transformation, above and below the benchmarks.
}
\author{
Vikas Yadav and Sunny Kumar
}
